@@ -3,11 +3,10 @@
   "version": "4.1.0",
   "private": true,
   "dependencies": {
-    "react": "^16.8.4",
-    "react-dom": "^16.8.4",
-    "react-ga": "^2.5.7",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "^5.0.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-ga": "^3.3.1",
+    "react-router-dom": "^6.3.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -27,6 +26,7 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "gh-pages": "^2.1.1"
+    "gh-pages": "^4.0.0",
+    "react-scripts": "^5.0.1"
   }
 }
